+++ examples/drift/lib/mappr.dart
@@ -1,5 +1,5 @@
 import 'package:auto_mappr_annotation/auto_mappr_annotation.dart';
-import 'package:auto_mappr_drift_example/db.dart';
+import 'package:examples_drift/db.dart';
 
 part 'mappr.g.dart';