+++ examples/injectable/lib/getit.dart
@@ -1,4 +1,4 @@
-import 'package:auto_mappr_injectable_example/getit.config.dart';
+import 'package:examples_injectable/getit.config.dart';
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';