+++ packages/auto_mappr/lib/src/builder/value_assignment_builder.dart
@@ -89,7 +89,7 @@ class ValueAssignmentBuilder {
           isOnNullable: sourceNullable,
         );
 
-    final defaultIterableValueExpression = targetType.defaultIterableExpression();
+    final defaultIterableValueExpression = targetType.defaultIterableExpression(config: mapperConfig);
 
     if (assignNestedObject) {
       return sourceIterableExpression
@@ -100,9 +100,9 @@ class ValueAssignmentBuilder {
             {},
             [
               refer(
-                targetIterableType.getDisplayString(
+                targetIterableType.getDisplayStringWithLibraryAlias(
                   withNullability: true,
-                  // classType: assignment.typeMapping.target,
+                  config: mapperConfig,
                 ),
               )
             ],
@@ -164,8 +164,8 @@ class ValueAssignmentBuilder {
 
     final defaultMapValueExpression = literalMap(
       {},
-      refer(targetKeyType.getDisplayString(withNullability: true)),
-      refer(targetValueType.getDisplayString(withNullability: true)),
+      refer(targetKeyType.getDisplayStringWithLibraryAlias(withNullability: true, config: mapperConfig)),
+      refer(targetValueType.getDisplayStringWithLibraryAlias(withNullability: true, config: mapperConfig)),
     );
 
     final assignNestedObjectKey = !targetKeyType.isPrimitiveType && (targetKeyType != sourceKeyType);
@@ -181,6 +181,7 @@ class ValueAssignmentBuilder {
       valueIsNullable: shouldRemoveNullsValue,
       keyType: sourceKeyType,
       valueType: sourceValueType,
+      config: mapperConfig,
     )
         .maybeCall(
       'map',
@@ -191,8 +192,8 @@ class ValueAssignmentBuilder {
         _callMapForMap(assignment),
       ],
       typeArguments: [
-        refer(targetKeyType.getDisplayString(withNullability: true)),
-        refer(targetValueType.getDisplayString(withNullability: true)),
+        refer(targetKeyType.getDisplayStringWithLibraryAlias(withNullability: true, config: mapperConfig)),
+        refer(targetValueType.getDisplayStringWithLibraryAlias(withNullability: true, config: mapperConfig)),
       ],
     )
         // When [sourceNullable], use default value.
@@ -221,7 +222,10 @@ class ValueAssignmentBuilder {
     );
 
     if (nestedMapping == null) {
-      final targetTypeName = target.getDisplayString(withNullability: true);
+      final targetTypeName = target.getDisplayStringWithLibraryAlias(
+        withNullability: true,
+        config: mapperConfig,
+      );
       final sourceName = assignment.sourceField?.getDisplayString(withNullability: true);
 
       if (target.nullabilitySuffix == NullabilitySuffix.question) {
@@ -292,10 +296,12 @@ class ValueAssignmentBuilder {
           ? MethodBuilderBase.constructNullableConvertMethodName(
               source: source,
               target: target,
+              config: mapperConfig,
             )
           : MethodBuilderBase.constructConvertMethodName(
               source: source,
               target: target,
+              config: mapperConfig,
             ),
     );
 
@@ -310,8 +316,8 @@ class ValueAssignmentBuilder {
             {},
             includeGenericTypes
                 ? [
-                    refer(source.getDisplayString(withNullability: true)),
-                    refer(target.getDisplayString(withNullability: true))
+                    refer(source.getDisplayStringWithLibraryAlias(withNullability: true, config: mapperConfig)),
+                    refer(target.getDisplayStringWithLibraryAlias(withNullability: true, config: mapperConfig))
                   ]
                 : [],
           );