+++ packages/auto_mappr/lib/src/extensions/dart_type_extension.dart
@@ -1,5 +1,7 @@
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
+import 'package:auto_mappr/src/extensions/element_extension.dart';
+import 'package:auto_mappr/src/models/auto_mappr_config.dart';
 import 'package:code_builder/code_builder.dart';
 import 'package:collection/collection.dart';
 
@@ -54,44 +56,98 @@ extension DartTypeExtension on DartType {
 
     final isSameExceptNullability = isSameName && isSameLibrary;
 
-    if (withNullability) {
+    if (!withNullability) {
       return isSameExceptNullability;
     }
 
     // Nullability matches.
-    final thisNullability = nullabilitySuffix == NullabilitySuffix.star;
-    final otherNullability = other.nullabilitySuffix == NullabilitySuffix.star;
+    final thisNullability = nullabilitySuffix == NullabilitySuffix.question;
+    final otherNullability = other.nullabilitySuffix == NullabilitySuffix.question;
     final isSameNullability = thisNullability == otherNullability;
 
     return isSameExceptNullability && isSameNullability;
   }
 
-  Expression defaultIterableExpression() {
+  Expression defaultIterableExpression({
+    required AutoMapprConfig config,
+  }) {
     final itemType = genericParameterTypeOrSelf;
 
-    if (isDartCoreList) {
-      return literalList([], refer(itemType.getDisplayString(withNullability: true)));
+    if (isDartCoreSet) {
+      return literalSet(
+        {},
+        refer(
+          itemType.getDisplayStringWithLibraryAlias(
+            config: config,
+            withNullability: true,
+          ),
+        ),
+      );
     }
 
-    // set
-    return literalSet({}, refer(itemType.getDisplayString(withNullability: true)));
+    return literalList(
+      [],
+      refer(
+        itemType.getDisplayStringWithLibraryAlias(
+          config: config,
+          withNullability: true,
+        ),
+      ),
+    );
   }
 
-  String toConvertMethodName({
-    required bool withNullability,
+  String getDisplayStringWithLibraryAlias({
+    required AutoMapprConfig config,
+    bool withNullability = false,
   }) {
-    final buffer = StringBuffer()..write(element!.name);
+    final alias = element!.getLibraryAlias(config: config);
+    final typeName = element!.name;
+    final buffer = StringBuffer('$alias$typeName');
 
     if (this is ParameterizedType && (this as ParameterizedType).typeArguments.isNotEmpty) {
       final arguments = (this as ParameterizedType)
           .typeArguments
-          .map<String>((argument) => argument.toConvertMethodName(withNullability: withNullability))
+          .map<String>(
+            (argument) => argument.getDisplayStringWithLibraryAlias(
+              withNullability: withNullability,
+              config: config,
+            ),
+          )
+          .join(',');
+      buffer.write('<$arguments>');
+    }
+
+    // Nullability
+    if (withNullability && nullabilitySuffix == NullabilitySuffix.question) {
+      buffer.write('?');
+    }
+
+    return buffer.toString();
+  }
+
+  String toConvertMethodName({
+    required bool withNullability,
+    required AutoMapprConfig config,
+  }) {
+    final alias = element!.getLibraryAlias(config: config, postfix: '_');
+    final typeName = element!.name;
+    final buffer = StringBuffer()..write('$alias$typeName');
+
+    if (this is ParameterizedType && (this as ParameterizedType).typeArguments.isNotEmpty) {
+      final arguments = (this as ParameterizedType)
+          .typeArguments
+          .map<String>(
+            (argument) => argument.toConvertMethodName(
+              withNullability: withNullability,
+              config: config,
+            ),
+          )
           .join(r'$');
       buffer.write('\$$arguments');
     }
 
     // Nullability
-    if (withNullability && nullabilitySuffix == NullabilitySuffix.star) {
+    if (withNullability && nullabilitySuffix == NullabilitySuffix.question) {
       buffer.write('?');
     }