+++ packages/auto_mappr/lib/src/extensions/list_extension.dart
@@ -1,0 +1,16 @@
+extension ListExtension<T> on List<T> {
+  List<T> get duplicates {
+    final dup = <T>[];
+    final buffer = <T>[];
+
+    for (final x in this) {
+      if (buffer.contains(x)) {
+        dup.add(x);
+      } else {
+        buffer.add(x);
+      }
+    }
+
+    return dup;
+  }
+}