+++ packages/auto_mappr/lib/src/generator/auto_mappr_generator.dart
@@ -1,13 +1,17 @@
 //ignore_for_file: avoid-dynamic
 
+import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:auto_mappr/src/builder/auto_mappr_builder.dart';
 import 'package:auto_mappr/src/extensions/dart_object_extension.dart';
+import 'package:auto_mappr/src/extensions/dart_type_extension.dart';
+import 'package:auto_mappr/src/extensions/list_extension.dart';
 import 'package:auto_mappr/src/models/models.dart';
 import 'package:auto_mappr_annotation/auto_mappr_annotation.dart';
 import 'package:build/build.dart';
 import 'package:code_builder/code_builder.dart';
+import 'package:collection/collection.dart';
 import 'package:source_gen/source_gen.dart';
 
 /// Code generator to generate implemented mapping classes.
@@ -22,13 +26,59 @@ class AutoMapprGenerator extends GeneratorForAnnotation<AutoMappr> {
       );
     }
 
+    final libraryUriToAlias = _getLibraryAliases(element: element);
+
+    final tmpConfig = AutoMapprConfig(
+      mappers: [],
+      availableMappingsMacroId: 'tmp',
+      libraryUriToAlias: libraryUriToAlias,
+    );
+
     final constant = annotation.objectValue;
     final mappersField = constant.getField('mappers')!;
     final mappersList = mappersField.toListValue()!;
-    final modulesExpression = constant.getField('modules')!.toCodeExpression();
+    final modulesExpression = constant.getField('modules')!.toCodeExpression(config: tmpConfig);
     final modulesList = constant.getField('modules')!.toListValue();
 
-    final mappers = mappersList.map((mapper) {
+    final mappers = _processMappers(
+      mappers: mappersList,
+      element: element,
+      config: tmpConfig,
+    );
+
+    final duplicates = mappers.duplicates;
+    if (duplicates.isNotEmpty) {
+      throw InvalidGenerationSourceError(
+        '@AutoMappr has configured duplicated mappings:\n\t${duplicates.map(
+              (e) => e.toStringWithLibraryAlias(
+                config: tmpConfig,
+              ),
+            ).join('\n\t')}',
+      );
+    }
+
+    final config = AutoMapprConfig(
+      mappers: mappers,
+      availableMappingsMacroId: element.library.identifier,
+      libraryUriToAlias: libraryUriToAlias,
+      modulesCode: modulesExpression,
+      modulesList: modulesList ?? [],
+    );
+
+    final builder = AutoMapprBuilder(mapperClassElement: element, config: config);
+
+    final output = builder.build();
+    final emitter = DartEmitter(orderDirectives: true, useNullSafetySyntax: true);
+
+    return '${output.accept(emitter)}';
+  }
+
+  List<TypeMapping> _processMappers({
+    required List<DartObject> mappers,
+    required ClassElement element,
+    required AutoMapprConfig config,
+  }) {
+    return mappers.map((mapper) {
       final mapperType = mapper.type! as ParameterizedType;
 
       final sourceType = mapperType.typeArguments.first;
@@ -36,21 +86,21 @@ class AutoMapprGenerator extends GeneratorForAnnotation<AutoMappr> {
 
       if (sourceType is! InterfaceType) {
         throw InvalidGenerationSourceError(
-          '${sourceType.getDisplayString(withNullability: true)} is not a class and cannot be mapped from',
+          '${sourceType.getDisplayStringWithLibraryAlias(config: config, withNullability: true)} is not a class and cannot be mapped from',
           element: element,
           todo: 'Use a class',
         );
       }
       if (targetType is! InterfaceType) {
         throw InvalidGenerationSourceError(
-          '${targetType.getDisplayString(withNullability: true)} is not a class and cannot be mapped to',
+          '${targetType.getDisplayStringWithLibraryAlias(config: config, withNullability: true)} is not a class and cannot be mapped to',
           element: element,
           todo: 'Use a class',
         );
       }
 
       final fields = mapper.getField('fields')?.toListValue();
-      final whenSourceIsNull = mapper.getField('whenSourceIsNull')?.toCodeExpression();
+      final whenSourceIsNull = mapper.getField('whenSourceIsNull')?.toCodeExpression(config: config);
       final constructor = mapper.getField('constructor')?.toStringValue();
 
       final fieldMappings = fields
@@ -59,8 +109,9 @@ class AutoMapprGenerator extends GeneratorForAnnotation<AutoMappr> {
               field: fieldMapping.getField('field')!.toStringValue()!,
               ignore: fieldMapping.getField('ignore')!.toBoolValue()!,
               from: fieldMapping.getField('from')!.toStringValue(),
-              customExpression: fieldMapping.getField('custom')!.toCodeExpression(passModelArgument: true),
-              whenNullExpression: fieldMapping.getField('whenNull')!.toCodeExpression(),
+              customExpression:
+                  fieldMapping.getField('custom')!.toCodeExpression(passModelArgument: true, config: config),
+              whenNullExpression: fieldMapping.getField('whenNull')!.toCodeExpression(config: config),
             ),
           )
           .toList();
@@ -73,43 +124,43 @@ class AutoMapprGenerator extends GeneratorForAnnotation<AutoMappr> {
         constructor: constructor,
       );
     }).toList();
-
-    final duplicates = mappers.duplicates;
-    if (duplicates.isNotEmpty) {
-      throw InvalidGenerationSourceError(
-        '@AutoMappr has configured duplicated mappings:\n\t${duplicates.join('\n\t')}',
-      );
-    }
-
-    final config = AutoMapprConfig(
-      mappers: mappers,
-      availableMappingsMacroId: element.library.identifier,
-      modulesCode: modulesExpression,
-      modulesList: modulesList ?? [],
-    );
-
-    final builder = AutoMapprBuilder(mapperClassElement: element, config: config);
-
-    final output = builder.build();
-    final emitter = DartEmitter(orderDirectives: true, useNullSafetySyntax: true);
-
-    return '${output.accept(emitter)}';
   }
-}
 
-extension ListEx<T> on List<T> {
-  List<T> get duplicates {
-    final dup = <T>[];
-    final buffer = <T>[];
+  Map<String, String> _getLibraryAliases({
+    required ClassElement element,
+  }) {
+    final libraryUriToAlias = <String, String>{};
 
-    for (final x in this) {
-      if (buffer.contains(x)) {
-        dup.add(x);
-      } else {
-        buffer.add(x);
-      }
+    final imports = element.library.libraryImports;
+    final aliases = imports.map((e) => e.prefix?.element.name).toList();
+    final uris = imports.map((e) => e.importedLibrary!.identifier).toList();
+
+    for (var i = 0; i < imports.length; i++) {
+      final currentAlias = aliases[i];
+      if (currentAlias == null) continue;
+
+      final importedLibrary = imports[i].importedLibrary!;
+      final exports = _getRecursiveLibraryExports(importedLibrary);
+
+      libraryUriToAlias.addAll({
+        // Current library.
+        uris[i]: currentAlias,
+        // It's exports.
+        for (final exported in exports) exported.identifier: currentAlias,
+      });
     }
 
-    return dup;
+    return libraryUriToAlias;
+  }
+
+  /// Recursively returns all exports (even nested) for [library].
+  Iterable<LibraryElement> _getRecursiveLibraryExports(LibraryElement library) {
+    final exports = library.libraryExports;
+    if (exports.isEmpty) return [];
+
+    return [
+      ...exports.map((e) => e.exportedLibrary!),
+      ...exports.map((e) => _getRecursiveLibraryExports(e.exportedLibrary!)).flattened,
+    ];
   }
 }