+++ packages/auto_mappr/test/integration/fixture/import_alias/import_alias_1b.dart
@@ -1,0 +1,26 @@
+import 'package:equatable/equatable.dart';
+
+export 'import_alias_1c.dart';
+
+class XxxDto {
+  final String name;
+  final int age;
+
+  const XxxDto({
+    required this.name,
+    required this.age,
+  });
+}
+
+class Xxx with EquatableMixin {
+  final String name;
+  final int age;
+
+  @override
+  List<Object?> get props => [name, age];
+
+  const Xxx({
+    required this.name,
+    required this.age,
+  });
+}