+++ packages/auto_mappr/CHANGELOG.md
@@ -1,5 +1,9 @@
 [//]: # (## Unreleased)
 
+## Unreleased
+- Add support for library aliases, so mapping now supports types from different libraries with the same name.
+It also work in cases when library alias's exports. [#62](https://github.com/netglade/auto_mappr/pull/62)
+
 ## 1.4.0
 - Add modules. [#67](https://github.com/netglade/auto_mappr/pull/67)
 - Split method builder into separate files. [#67](https://github.com/netglade/auto_mappr/pull/67)